detail page connected

+++ lib/Screens/adminscreen/admin_gridview.dart
@@ -1,15 +1,15 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:recipely/adminscreen/admin_add.dart';
-import 'package:recipely/adminscreen/admin_update.dart';
+import 'package:recipely/Screens/adminscreen/admin_add.dart';
+import 'package:recipely/Screens/adminscreen/admin_update.dart';
 import 'package:recipely/datas/hive_db.dart';
 import 'package:recipely/models/model_recipe.dart';
 
 class Admingridview extends StatefulWidget {
   const Admingridview({
-    super.key,
-  });
+    Key? key,
+  }) : super(key: key);
 
   @override
   State<Admingridview> createState() => _AdmingridviewState();
@@ -29,19 +29,20 @@ class _AdmingridviewState extends State<Admingridview> {
   @override
   void initState() {
     fetchdata();
-
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          centerTitle: true,
-          backgroundColor: const Color(0xFF1EDEC7),
-          title: const Text("Added List"),
-        ),
-        body: GridView.builder(
+      appBar: AppBar(
+        centerTitle: true,
+        backgroundColor: const Color(0xFF1EDEC7),
+        title: const Text("Added List"),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(8.0),
+        child: GridView.builder(
           gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
             childAspectRatio: 1 / 1.5,
             crossAxisSpacing: 10,
@@ -50,7 +51,6 @@ class _AdmingridviewState extends State<Admingridview> {
           ),
           itemCount: datalist.length,
           itemBuilder: (context, index) {
-            // Check if the file exists before loading the image
             File imageFile = File(datalist[index].photo);
             String timeKey = datalist[index].timeKey ?? '';
             return GestureDetector(
@@ -61,30 +61,26 @@ class _AdmingridviewState extends State<Admingridview> {
                 children: [
                   InkWell(
                     onTap: () {
-                      // reciepeOfIndex = datalist[index];
                       Navigator.of(context).push(MaterialPageRoute(
                         builder: (context) => Updatescrren(
                             reciepeOfIndexForEditng: datalist[index]),
                       ));
                     },
                     child: ClipRRect(
-                      borderRadius: BorderRadius.circular(
-                          10.0), // Adjust the radius as needed
+                      borderRadius: BorderRadius.circular(10.0),
                       child: Container(
                         decoration: BoxDecoration(
                           boxShadow: [
                             BoxShadow(
-                              color:
-                                  Colors.black.withOpacity(0.2), // Shadow color
-                              spreadRadius: 2, // Spread radius
-                              blurRadius: 5, // Blur radius
-                              offset: const Offset(
-                                  0, 3), // Offset in the direction of shadow
+                              color: Colors.black.withOpacity(0.2),
+                              spreadRadius: 2,
+                              blurRadius: 5,
+                              offset: const Offset(0, 3),
                             ),
                           ],
                         ),
                         child: AspectRatio(
-                          aspectRatio: 1 / 1.5, // Set the desired aspect ratio
+                          aspectRatio: 1 / 1.5,
                           child: Image.file(
                             imageFile,
                             fit: BoxFit.cover,
@@ -99,9 +95,36 @@ class _AdmingridviewState extends State<Admingridview> {
                     child: IconButton(
                       icon: const Icon(Icons.delete, color: Colors.white),
                       onPressed: () async {
-                        await deleteRecipe(timeKey);
-                        print('Deleting recipe with key: $timeKey');
-                        fetchdata();
+                        bool confirmDelete = await showDialog(
+                          context: context,
+                          builder: (BuildContext context) {
+                            return AlertDialog(
+                              title: const Text('Confirm Delete'),
+                              content: const Text(
+                                  'Are you sure you want to delete this recipe?'),
+                              actions: [
+                                TextButton(
+                                  onPressed: () {
+                                    Navigator.of(context).pop(false);
+                                  },
+                                  child: const Text('No'),
+                                ),
+                                TextButton(
+                                  onPressed: () {
+                                    Navigator.of(context).pop(true);
+                                  },
+                                  child: const Text('Yes'),
+                                ),
+                              ],
+                            );
+                          },
+                        );
+
+                        if (confirmDelete == true) {
+                          await deleteRecipe(timeKey);
+                          print('Deleting recipe with key: $timeKey');
+                          fetchdata();
+                        }
                       },
                     ),
                   ),
@@ -110,16 +133,17 @@ class _AdmingridviewState extends State<Admingridview> {
             );
           },
         ),
-        floatingActionButton: FloatingActionButton(
-          onPressed: () {
-            Navigator.of(context).push(MaterialPageRoute(
-              builder: (context) => const Addingscreen(),
-            ));
-          },
-          backgroundColor: const Color(0xFF1EDEC7),
-          child: const Icon(Icons.add),
-        ),
-        floatingActionButtonLocation:
-            FloatingActionButtonLocation.centerDocked);
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          Navigator.of(context).push(MaterialPageRoute(
+            builder: (context) => const Addingscreen(),
+          ));
+        },
+        backgroundColor: const Color(0xFF1EDEC7),
+        child: const Icon(Icons.add),
+      ),
+      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
+    );
   }
 }
